% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_linear_reduction.R
\name{clusters_seurat}
\alias{clusters_seurat}
\title{Non linear dimensionality reduction for scRNAseq}
\usage{
clusters_seurat(input)
}
\arguments{
\item{input}{is the output of scaling (for function clusters_seurat()). Use the output of the clusters function as input for t-SNE and UMAP functions}
}
\description{
t-SNE and UMAP functions are included in this package
}
\examples{
tsne_seurat(after_scaling.rds)
}
