% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_linear_reduction.R
\name{tsne_seurat}
\alias{tsne_seurat}
\title{Non linear dimensionality reduction for scRNAseq}
\usage{
tsne_seurat(input)
}
\arguments{
\item{input}{is the output of RunPCA}
}
\description{
t-SNE and UMAP functions are included in this package
}
\examples{
tsne_seurat(afterclustering.rds)
}
